// src/types/infosType.d.ts
type nullNumber = null | number
export interface FactorInfoType {
    customer_income_average: nullNumber,
factor_amount_average : nullNumber,
factor_amount_product_average : nullNumber,
factor_commisions : nullNumber,
factor_product_average : nullNumber,
factor_product_count : nullNumber,
factor_product_weight : nullNumber,
factor_rows_average : nullNumber,
gross_new_users : nullNumber,
gross_rejected : nullNumber,
gross_sale : nullNumber,
gross_sale_per_country : {labels: [], data: []}
gross_sale_per_customer_gender : {labels: [], data: []}
gross_sale_per_product_brand : {labels: [], data: []}
gross_sale_per_product_category : {labels: [], data: []}
gross_sale_per_sales_channel : {labels: [], data: []}
new_users_count : nullNumber,
old_users_count : nullNumber,
percent_rejected_amount : nullNumber,
percent_rejected_count : nullNumber,
percent_users : nullNumber,
pure_factor_counts : nullNumber,
pure_gross_sale : nullNumber,
pure_product_count : nullNumber,
pure_product_weight : nullNumber,
pure_sale : nullNumber,
rejected_factor_count : nullNumber,
rejected_product_count : nullNumber,
rejedted_product_weight : nullNumber,
sale_factor_count : nullNumber,
sum_discount : nullNumber,
users_count : nullNumber,
}
export interface ProductInfoType {
    customer_income_average : nullNumber,
    factor_amount_average : nullNumber,
    factor_amount_product_average : nullNumber,
    factor_commisions : nullNumber,
    factor_product_average : nullNumber,
    factor_product_count : nullNumber,
    factor_product_weight : nullNumber,
    factor_rows_average  : nullNumber,
    gross_new_users : nullNumber,
    gross_rejected : nullNumber,
    gross_sale : nullNumber,
    new_users_count : nullNumber,
    old_users_count : nullNumber,
    percent_rejected_amount : nullNumber,
    percent_rejected_count : nullNumber,
    percent_users : nullNumber,
    pure_factor_counts : nullNumber,
    pure_gross_sale : nullNumber,
    pure_product_count : nullNumber,
    pure_product_weight : nullNumber,
    pure_sale : nullNumber,
    rejected_factor_count : nullNumber,
    rejected_product_count : nullNumber,
    rejedted_product_weight : nullNumber,
    sale_factor_count : nullNumber,
    sum_discount : nullNumber,
    users_count : nullNumber,
}